[feat] 강의 상세페이지 탭 메뉴 클릭시 페이지 이동따라 GET 분기처리
